区域变更12位&汇总车辆覆盖乡镇

+++ obd-core/src/main/java/com/wkhmedical/dto/ObdCarDTO.java
@@ -67,6 +67,6 @@ public class ObdCarDTO implements Serializable {
 	private String carName;
 
 	@ApiModelProperty(value = "区县ID")
-	private Integer areaId;
+	private Long areaId;
 
 }
