+++ obd-web/src/main/java/com/wkhmedical/web/api/BindUserController.java
@@ -96,7 +96,7 @@ public class BindUserController {
 
 	@ApiOperation(value = "获取当前区域的默认负责人")
 	@GetMapping("/get.default")
-	public List<BindUser> getBindUserDefault(@ApiParam(value = "区县ID", required = true) @RequestParam Integer areaId,
+	public List<BindUser> getBindUserDefault(@ApiParam(value = "区县ID", required = true) @RequestParam Long areaId,
 			@ApiParam(value = "车辆ID") @RequestParam String eid) {
 		return bindUserService.getBindUserDefault(areaId, eid);
 	}
